+++ install.py
@@ -9,7 +9,7 @@ import secrets
 
 from weecfg.extension import ExtensionInstaller
 
-VERSION = '0.12.2'
+VERSION = '0.12.3'
 
 
 def loader():